Add initialize to the plugin class
