+++ plugins/kernel_v2/config/vm.rb
@@ -724,9 +724,10 @@ module VagrantPlugins
         errors = _detected_errors
 
         if @allow_fstab_modification == UNSET_VALUE
-          machine.synced_folders.each do |f|
-            if f.capability?(:default_fstab_modification)
-              if f.capability(:default_fstab_modification) == false
+          # require "pry-byebug"; binding.pry
+          machine.synced_folder_types.each do |type|
+            if type.capability?(:default_fstab_modification)
+              if type.capability(:default_fstab_modification) == false
                 @allow_fstab_modification = false
                 break
               end
